Split out deps script to lib files
